Moving the _get_or_generate_backend_id
The backend_id is required in multiple places, so
getting or generating it should be done previously.
